pyproject.toml
[tool.poetry]
name = "format-django-backend"
version = "0.1.0"
description = ""
authors = ["phuyalbish <phuyalbish@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
markdown = "^3.7"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
gunicorn = "^23.0.0"
whitenoise = "^6.8.2"
psycopg = "^3.2.4"
django-admin-interface = "^0.29.4"
django-colorfield = "^0.12.0"
psycopg2-binary = "^2.9.10"
django-tinymce = "^4.1.0"
mysqlclient = "^2.2.7"
pylint = "^3.3.6"
autoflake = "^2.3.1"
cloudinary = "^1.44.0"
requests = "^2.32.3"
django-extensions = "^4.1"
django-jazzmin = "^3.0.1"
werkzeug = "^3.1.5"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pylint = "^3.3.6"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
